+++ activemodel/lib/active_model/type/integer.rb
@@ -3,15 +3,6 @@ module ActiveModel
     class Integer < Value # :nodoc:
       include Helpers::Numeric
 
-      # Column storage size in bytes.
-      # 4 bytes means a MySQL int or Postgres integer as opposed to smallint etc.
-      DEFAULT_LIMIT = 4
-
-      def initialize(*)
-        super
-        @range = min_value...max_value
-      end
-
       def type
         :integer
       end
@@ -22,16 +13,8 @@ module ActiveModel
       end
 
       def serialize(value)
-        result = cast(value)
-        if result
-          ensure_in_range(result)
-        end
-        result
+        cast(value)
       end
-
-      protected
-
-        attr_reader :range
 
       private
 
@@ -42,24 +25,6 @@ module ActiveModel
           else
             value.to_i rescue nil
           end
-        end
-
-        def ensure_in_range(value)
-          unless range.cover?(value)
-            raise ActiveModel::RangeError, "#{value} is out of range for #{self.class} with limit #{_limit}"
-          end
-        end
-
-        def max_value
-          1 << (_limit * 8 - 1) # 8 bits per byte with one bit for sign
-        end
-
-        def min_value
-          -max_value
-        end
-
-        def _limit
-          limit || DEFAULT_LIMIT
         end
     end
   end