+++ activemodel/lib/active_model/type/unsigned_integer.rb
@@ -1,15 +1,15 @@
+require "active_model/type/integer"
+
 module ActiveModel
   module Type
     class UnsignedInteger < Integer # :nodoc:
-      private
-
-        def max_value
-          super * 2
-        end
-
-        def min_value
-          0
-        end
+      def initialize(*)
+        ActiveSupport::Deprecation.warn(<<-MSG.squish)
+          #{self.class} is deprecated and will be removed in Rails 5.2.
+          Please use #{self.class.superclass} instead.
+        MSG
+        super
+      end
     end
   end
 end