+++ activerecord/lib/active_record/connection_adapters/abstract_mysql_adapter.rb
@@ -693,13 +693,7 @@ module ActiveRecord
         end
 
         def register_integer_type(mapping, key, options) # :nodoc:
-          mapping.register_type(key) do |sql_type|
-            if /\bunsigned\z/ === sql_type
-              Type::UnsignedInteger.new(options)
-            else
-              Type::Integer.new(options)
-            end
-          end
+          mapping.register_type(key) { Type::Integer.new(options) }
         end
 
         def extract_precision(sql_type)
@@ -736,6 +730,7 @@ module ActiveRecord
         ER_DUP_ENTRY            = 1062
         ER_NO_REFERENCED_ROW_2  = 1452
         ER_DATA_TOO_LONG        = 1406
+        ER_OUT_OF_RANGE         = 1264
         ER_LOCK_DEADLOCK        = 1213
 
         def translate_exception(exception, message)
@@ -746,6 +741,8 @@ module ActiveRecord
             InvalidForeignKey.new(message)
           when ER_DATA_TOO_LONG
             ValueTooLong.new(message)
+          when ER_OUT_OF_RANGE
+            RangeError.new(message)
           when ER_LOCK_DEADLOCK
             Deadlocked.new(message)
           else