+++ activerecord/test/cases/adapter_test.rb
@@ -220,6 +220,14 @@ module ActiveRecord
 
         assert_not_nil error.cause
       end
+
+      def test_numeric_value_out_of_ranges_are_translated_to_specific_exception
+        error = assert_raises(ActiveRecord::RangeError) do
+          Book.create(author_id: 2147483648)
+        end
+
+        assert_not_nil error.cause
+      end
     end
 
     def test_disable_referential_integrity