+++ activerecord/test/cases/adapters/mysql2/unsigned_type_test.rb
@@ -28,16 +28,16 @@ class Mysql2UnsignedTypeTest < ActiveRecord::Mysql2TestCase
   end
 
   test "minus value is out of range" do
-    assert_raise(ActiveModel::RangeError) do
+    assert_raise(ActiveRecord::RangeError) do
       UnsignedType.create(unsigned_integer: -10)
     end
-    assert_raise(ActiveModel::RangeError) do
+    assert_raise(ActiveRecord::RangeError) do
       UnsignedType.create(unsigned_bigint: -10)
     end
-    assert_raise(ActiveRecord::StatementInvalid) do
+    assert_raise(ActiveRecord::RangeError) do
       UnsignedType.create(unsigned_float: -10.0)
     end
-    assert_raise(ActiveRecord::StatementInvalid) do
+    assert_raise(ActiveRecord::RangeError) do
       UnsignedType.create(unsigned_decimal: -10.0)
     end
   end