+++ activerecord/test/cases/attributes_test.rb
@@ -162,16 +162,16 @@ module ActiveRecord
     end
 
     if current_adapter?(:PostgreSQLAdapter)
+      OID = ConnectionAdapters::PostgreSQL::OID
+
       test "array types can be specified" do
         klass = Class.new(OverloadedType) do
           attribute :my_array, :string, limit: 50, array: true
           attribute :my_int_array, :integer, array: true
         end
 
-        string_array = ConnectionAdapters::PostgreSQL::OID::Array.new(
-          Type::String.new(limit: 50))
-        int_array = ConnectionAdapters::PostgreSQL::OID::Array.new(
-          Type::Integer.new)
+        string_array = OID::Array.new(Type::String.new(limit: 50))
+        int_array = OID::Array.new(OID::Integer.new)
         assert_not_equal string_array, int_array
         assert_equal string_array, klass.type_for_attribute("my_array")
         assert_equal int_array, klass.type_for_attribute("my_int_array")
@@ -183,10 +183,8 @@ module ActiveRecord
           attribute :my_int_range, :integer, range: true
         end
 
-        string_range = ConnectionAdapters::PostgreSQL::OID::Range.new(
-          Type::String.new(limit: 50))
-        int_range = ConnectionAdapters::PostgreSQL::OID::Range.new(
-          Type::Integer.new)
+        string_range = OID::Range.new(Type::String.new(limit: 50))
+        int_range = OID::Range.new(OID::Integer.new)
         assert_not_equal string_range, int_range
         assert_equal string_range, klass.type_for_attribute("my_range")
         assert_equal int_range, klass.type_for_attribute("my_int_range")