+++ activemodel/CHANGELOG.md
@@ -1,2 +1,10 @@
+*   Deprecate `Type::UnsignedInteger`. Use `Type::Integer` instead.
+
+    *Ryuta Kamizono*
+
+*   Deprecate `Type::BigInteger`. Use `Type::Integer` instead.
+
+    *Ryuta Kamizono*
+
 
 Please check [5-0-stable](https://github.com/rails/rails/blob/5-0-stable/activemodel/CHANGELOG.md) for previous changes.